feat(core)!: the contextual tool set moves to rig_core::tool; ToolCatalog; rig_tool resolves through rig-core
Tool, ToolEmbedding, ErasedTool, DynamicTool, RegisteredTool, ToolDispatch,
dispatch_tool, tool_definition and ToolSet move from rig_agent::tool to
rig_core::tool::contextual (re-exported at rig_core::tool::*), together with
their fixture-free unit tests. A systems driver can now hold a ToolSet in
shared state and dispatch by name with rig-core alone.

rig_core::tool::ToolCatalog is the old ToolRegistrySnapshot minus retrieval:
definitions plus pinned dispatch by name (execute/dispatch/retain_names/
take_definitions), built by ToolSet::catalog() (always-exposed tools,
registration order) or ToolCatalog::from_registered(ordered map). rig-agent
keeps ToolServer/ToolServerHandle (retrieval indexes, managed remote tool
sources, the per-turn snapshot) over these types; ToolRegistrySnapshot is
now `pub type ToolRegistrySnapshot = ToolCatalog`. ToolSet gains the reads
the registry needed instead of reaching into its map (names, len, is_empty,
get, always_exposed_names, add_retrievable_tools, move_to_end, catalog).
Pinned Send + Sync + 'static in rig-core: ToolSet, ToolCatalog, DynamicTool.

rig-derive evidence (crates/rig-derive/tests/dependency_rename.rs): the
macro never hardcoded crate paths — CrateRefs::resolve() picks the core
root from rig-core / rig::core / rig_agent::core — so the #[rig_tool]
expansion now emits rig_core::tool::{Tool, ToolContext} for contextual
tools as well, the "contextual tools require rig/rig-agent" diagnostic is
gone, rig-core's root is a recognised ToolContext path root, and the
core_only_contextual fixture flips from must-fail to must-compile.

The rig facade exports the rig-core tool types unconditionally at
rig::tool::*; rig::tool::server stays behind the agent feature. All old
rig_agent::tool::* paths keep working as re-exports.
