+++ crates/rig-derive/tests/fixtures/core_only_contextual/src/main.rs
@@ -1,15 +1,17 @@
-// A contextual tool in a crate that depends only on `rig-core` must fail with
-// a targeted diagnostic, not an unresolved `::rig_agent` path.
+// A contextual tool in a crate that depends only on `rig-core` (under a
+// Cargo rename) compiles: `Tool` and `ToolContext` are rig-core's, so no
+// runtime crate is needed to author one.
 use rig_macros::rig_tool;
-
-struct LocalContext;
 
 #[rig_tool]
 fn needs_runtime(
-    #[rig(context)] _context: &mut LocalContext,
+    #[rig(context)] _context: &mut core_runtime::tool::ToolContext,
     value: String,
 ) -> Result<String, core_runtime::tool::ToolExecutionError> {
     Ok(value)
 }
 
-fn main() {}
+fn main() {
+    let tool = NeedsRuntime;
+    let _definition = core_runtime::tool::tool_definition(&tool);
+}
